Block theme: Only show pattern thumbnail helper text for screen reader users, hide image on error
This removes the flash of text before the loader kicks in. This text is only used to ensure the link has an accessible name while the image is loading or broken. Once loaded, the image's alt serves this function. If there is an error, the (broken) image should not be visible, and the span provides the accessible text.

See https://github.com/WordPress/pattern-directory/pull/655#issuecomment-2055119364